+++ dom/tsdown.config.ts
@@ -1,0 +1,15 @@
+import { defineConfig } from 'tsdown';
+
+export default defineConfig({
+  entry: ['./src/index.ts'],
+  platform: 'neutral',
+  dts: true,
+  external: [
+    'react',
+    'react-dom',
+    'wy-helper',
+    'wy-dom-helper',
+    'wy-react-helper',
+  ],
+  format: ['esm', 'cjs'],
+});