+++ native/tsdown.config.ts
@@ -1,0 +1,16 @@
+import { defineConfig } from 'tsdown';
+
+export default defineConfig({
+  entry: ['./src/index.ts'],
+  platform: 'neutral',
+  dts: true,
+  external: [
+    'react',
+    'react-native',
+    'react-native-gesture-handler',
+    'react-native-reanimated',
+    'wy-helper',
+    'wy-react-helper',
+  ],
+  format: ['cjs', 'esm'],
+});